fix(docs): align uncertainty docs and physics headings
Align temperature sweep metadata and documentation with the current uncertainty behavior, including integer bootstrap_resamples metadata, single-seed entropy uncertainty fallback, and tau_int plot availability wording.

Fix the misplaced Bibliography heading in PHYSICS.md and docs/source/physics.md so the statistical uncertainty section appears in the correct order.

Validation: pytest -q; ruff check .; mypy --no-incremental --explicit-package-bases models/ utils/ scripts/; pre-commit run --all-files --hook-stage pre-push

+++ scripts/xy/temperature_sweep.py
@@ -557,7 +557,7 @@ def main() -> None:
         uncertainty_method=UNCERTAINTY_METHOD_BLOCKING,
         confidence_level=float(args.confidence_level),
         n_seeds=n_seeds,
-        bootstrap_resamples=float(args.entropy_bootstrap_resamples),
+        bootstrap_resamples=int(args.entropy_bootstrap_resamples),
         nan_or_undefined_count=float(np.isnan(tau_int_arr).sum()),
     )
